# api/openapi/平台接口规范.yaml
openapi: 3.1.0
info:
  title: 基础能力平台 API
  version: 0.1.0-draft
  description: |
    基础能力平台的前后端接口草案。实现基线为 Go 模块化单体、MySQL、JWT Cookie 会话，
    所有时间均为 UTC RFC 3339 格式。本文件与 docs/前后端接口契约.md 共同作为联调契约；
    表结构和领域约束以 docs/核心模型与MySQL设计.md、docs/系统架构.md 为准。
servers:
  - url: /api/v1
    description: 平台 API 相对地址
...
security:
  - cookieAuth: []
...
paths:
...
  /oauth2/token:
    servers:
      - url: /
        description: OAuth 2.0 根路径端点，不继承 /api/v1 前缀。
    post:
      tags: [OIDC, Application Registry]
      summary: 签发或轮换 OAuth/OIDC 令牌
      description: >-
        支持 authorization_code、refresh_token 与 client_credentials。authorization_code 和 refresh_token：
        confidential 客户端使用 HTTP Basic（client_secret_basic），public 客户端仅以表单 client_id 认证；
        不接受 client_secret 表单字段。client_credentials 必须使用 HTTP Basic。授权码兑换必须提供精确 redirect_uri，
        且使用 PKCE 的授权码必须提供匹配的 code_verifier。Refresh Token 每次成功使用均会轮换。
      operationId: issueOAuthToken
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type: { type: string, enum: [authorization_code, refresh_token, client_credentials] }
                client_id: { type: string, maxLength: 255, description: 仅 public authorization_code / refresh_token 客户端可提供。 }
                code: { type: string, description: grant_type=authorization_code 时必填。 }
                redirect_uri: { type: string, format: uri, description: grant_type=authorization_code 时必须与授权请求完全一致。 }
                code_verifier: { type: string, minLength: 43, maxLength: 128, description: 使用 PKCE 的 authorization_code 时必填。 }
                refresh_token: { type: string, description: grant_type=refresh_token 时必填。 }
                scope: { type: string, description: client_credentials 时为申请范围；授权码和刷新授权不会通过此字段扩大权限。 }
      responses:
        '200':
          description: OAuth 2.0 令牌响应；不使用平台通用响应信封。
          headers:
            Cache-Control: { schema: { type: string, example: no-store } }
            Pragma: { schema: { type: string, example: no-cache } }
          content:
            application/json: { schema: { $ref: '#/components/schemas/OAuthTokenResponse' } }
        '400':
          description: 无效请求、已消费/失效授权码、无效 Refresh Token、PKCE 校验失败、未授权授权类型或不允许的范围。
          content:
            application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } }
        '401':
          description: 客户端认证失败。
          headers:
            WWW-Authenticate: { schema: { type: string, example: 'Basic realm="oauth2/token"' } }
          content:
            application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } }
        '500':
          description: 令牌签发内部错误。
          content:
            application/json: { schema: { $ref: '#/components/schemas/OAuthErrorResponse' } }
...
  /oauth2/logout:
    servers:
      - url: /
        description: OpenID Connect 根路径端点，不继承 /api/v1 前缀。
    get:
      tags: [OIDC]
      summary: RP-initiated Logout
      description: >-
        尝试撤销当前平台 Cookie 会话并清理 Cookie。仅当 post_logout_redirect_uri 是客户端精确登记的回调地址，
        且提供了有效 id_token_hint 或相同客户端的 client_id 时才外部重定向；其他情况回到平台登录页。
      operationId: getOIDCLogout
      security: []
      parameters:
        - { name: id_token_hint, in: query, schema: { type: string } }
        - { name: client_id, in: query, schema: { type: string, maxLength: 255 } }
        - { name: post_logout_redirect_uri, in: query, schema: { type: string, format: uri, maxLength: 2048 } }
        - { name: state, in: query, schema: { type: string, maxLength: 2048 } }
      responses:
        '302':
          description: 已完成退出；重定向到验证后的客户端地址或平台登录页。
          headers:
            Location: { schema: { type: string, format: uri } }
            Set-Cookie: { schema: { type: string } }
    post:
      tags: [OIDC]
      summary: RP-initiated Logout
      description: 与 GET 相同；参数可通过 application/x-www-form-urlencoded 提交。
      operationId: postOIDCLogout
      security: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                id_token_hint: { type: string }
                client_id: { type: string, maxLength: 255 }
                post_logout_redirect_uri: { type: string, format: uri, maxLength: 2048 }
                state: { type: string, maxLength: 2048 }
      responses:
        '303':
          description: 已完成退出；重定向到验证后的客户端地址或平台登录页。
          headers:
            Location: { schema: { type: string, format: uri } }
            Set-Cookie: { schema: { type: string } }
...
  /auth/logout:
    post:
      tags: [Auth]
      summary: 退出当前会话
      operationId: logout
      x-permission: authenticated
      x-audit-action: auth.logout
      responses:
        '200':
          description: 当前 iam_session 已撤销且 Cookie 已清理。
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json: { schema: { $ref: '#/components/schemas/EmptyEnvelope' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /observability/logs:
    get:
      tags: [Observability]
      summary: 查询统一运行日志
      operationId: listRuntimeLogs
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: service, in: query, schema: { type: string } }, { name: level, in: query, schema: { type: string } }, { name: trace_id, in: query, schema: { type: string } }, { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } } ]
      responses: { '200': { description: 返回当前进程有界存储中的运行日志样本。 } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: bp_session
      description: JWT 会话 Cookie。实际名称由 IAM_SESSION_COOKIE_NAME 配置，默认 bp_session。
    applicationAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 由 /oauth2/token 使用 Client Credentials 签发的接入应用 JWT。接收端校验 Ed25519（EdDSA）签名、issuer、audience、token_use=application 与时间声明，并以数据库当前启用的 OAuth Client、Application、Environment 和 scope 复核令牌绑定；审计上报必须具备 audit.ingest 范围。
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 本 OIDC 发行方在 Authorization Code 或 Refresh Token 授权中签发的终端用户 Access Token。
    clientCredentialsBasic:
      type: http
      scheme: basic
      description: OAuth 令牌端点的 client_secret_basic 客户端认证。
...
  responses:
    Unauthenticated:
      description: 未登录、会话已过期或应用凭据无效。
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
    ULID: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$', examples: ['01J123456789ABCDEFGHJKMNPQ'] }
...
    Envelope:
      type: object
      required: [code, message, request_id, data]
      properties:
        code: { type: string, const: OK }
        message: { type: string, examples: ['操作成功'] }
        request_id: { $ref: '#/components/schemas/ULID' }
        data: {}
    EmptyEnvelope:
      allOf:
        - $ref: '#/components/schemas/Envelope'
        - type: object
          properties: { data: { type: object, additionalProperties: false } }
    ErrorEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          examples: [AUTH_UNAUTHENTICATED, AUTH_FORBIDDEN, IAM_VERSION_CONFLICT, PLATFORM_CAPABILITY_NOT_ENABLED]
        message: { type: string }
        request_id: { $ref: '#/components/schemas/ULID' }
        details: { type: object, additionalProperties: true }
...
    OAuthTokenResponse:
      type: object
      required: [access_token, token_type, expires_in, scope]
      properties:
        access_token: { type: string, description: 短期 Bearer JWT；Client Credentials 与 OIDC 终端用户令牌使用不同签发策略。 }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer, minimum: 1, description: Access Token 剩余有效期（秒）。 }
        scope: { type: string, description: 空格分隔的实际签发范围。 }
        refresh_token: { type: string, description: authorization_code 或 refresh_token 成功时返回；仅显示一次，服务端仅保存 SHA-256 摘要。 }
        id_token: { type: string, description: OpenID Connect Authorization Code 或 Refresh Token 成功时返回的 ID Token。 }
    OAuthErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string, enum: [invalid_request, invalid_grant, invalid_token, unsupported_grant_type, invalid_client, unauthorized_client, invalid_scope, server_error] }
...